requirements.txt
awscli==1.6.2
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
chainer==1.8.1
colorama==0.2.5
configparser==3.3.0.post2
cycler==0.10.0
dask==0.8.2
decorator==3.4.0
docutils==0.12
entrypoints==0.2.1
enum34==1.1.4
filelock==2.0.6
Flask==0.10.1
functools32==3.2.3.post2
genson==0.1.0
h5py==2.6.0
idna==2.1
ipaddress==1.0.16
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==5.1.3
itsdangerous==0.24
Jinja2==2.8
jmespath==0.5.0
jsonschema==2.5.1
jupyter-client==4.2.2
jupyter-core==4.1.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
multipledispatch==0.4.8
nbconvert==4.2.0
nbformat==4.0.1
networkx==1.11
nose==1.3.7
notebook==4.2.0
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==2.6.0
prettytable==0.7.2
protobuf==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.4
pyzmq==15.2.0
requests==2.9.1
rsa==3.1.2
scikit-image==0.12.3
scipy==0.17.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sseclient==0.0.8
terminado==0.6
toolz==0.7.4
tornado==4.1
traitlets==4.2.1
Werkzeug==0.11.9
widgetsnbextension==1.2.2
